+++ CHANGELOG.md
@@ -6,8 +6,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- add SwiftUI wrapper `StepsView`.
+
 ### Fixed
 - calculate `intrinsicContentSize`.
+- make SwiftUI frame correctly respond to `.fixedSize(horizontal: false, vertical: true)`.
 
 ## [1.5.7] - 2020-06-27